+++ test/macro.spec.js
@@ -186,19 +186,197 @@ test.group('Macroable', (group) => {
     assert.plan(2)
 
     macroable.addMacro('note', function (content, props, { eat }) {
-      assert.deepEqual(props, { title: ['hello world'], color: ['grey'] })
+      assert.deepEqual(props, {
+        title: 'hello world',
+        color: 'grey'
+      })
     })
 
     const template = dedent`
     Hello world!
 
-    [note title=hello world, color=grey]
+    [note title="hello world", color="grey"]
     Hey dude
     [/note]
     `
     const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
     assert.equal(result.contents, dedent`
       <p>Hello world!</p>\n
+    `)
+  })
+
+  test('do not trim macro content', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(2)
+
+    macroable.addMacro('note', function (content, props, { eat }) {
+      assert.equal(content, '  Hey dude')
+    })
+
+    const template = dedent`
+    Hello world!
+
+    [note title=hello world, color=grey]
+      Hey dude
+    [/note]
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+    assert.equal(result.contents, dedent`
+      <p>Hello world!</p>\n
+    `)
+  })
+
+  test('do not mess with anchor tags', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(1)
+
+    macroable.addMacro('note', function () {
+      throw new Error('Never expected to be called')
+    })
+
+    const template = dedent`
+    Hello world!
+
+    [note](hello)
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+    assert.equal(result.contents, dedent`
+      <p>Hello world!</p>
+      <p><a href="hello">note</a></p>\n
+    `)
+  })
+
+  test('allow macros inside li', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(1)
+
+    macroable.addMacro('note', function (content, props, { eat }) {
+      return {
+        type: 'text',
+        value: content
+      }
+    })
+
+    const template = dedent`
+    - List item
+
+        [note]
+        Hey dude
+        [/note]
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+    assert.equal(result.contents, dedent`
+      <ul>
+      <li>
+      <p>List item</p>
+      Hey dude
+      </li>
+      </ul>\n
+    `)
+  })
+
+  test('allow macros inside nested li', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(1)
+
+    macroable.addMacro('note', function (content, props, { eat }) {
+      return {
+        type: 'text',
+        value: content
+      }
+    })
+
+    const template = dedent`
+    - List item 1
+      - List item 1.1
+
+        [note]
+        Hey dude
+        [/note]
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+    assert.equal(result.contents, dedent`
+      <ul>
+      <li>
+      <p>List item 1</p>
+      <ul>
+      <li>
+      <p>List item 1.1</p>
+      Hey dude
+      </li>
+      </ul>
+      </li>
+      </ul>\n
+    `)
+  })
+
+  test('allow inline macros', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(1)
+
+    macroable.addMacro('codepen', function (props, { eat }) {
+      return {
+        type: 'text',
+        value: props.src
+      }
+    }, true)
+
+    const template = dedent`
+    Here is a pen
+
+    [codepen src=http://facebook.com]
+
+    Content after pen
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+
+    assert.equal(result.contents, dedent`
+    <p>Here is a pen</p>
+    http://facebook.com
+    <p>Content after pen</p>\n
+    `)
+  })
+
+  test('work fine when block macro is the last line', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(3)
+
+    macroable.addMacro('alert', function (content, props, { eat }) {
+      assert.equal(content, 'hello')
+      assert.deepEqual(props, {})
+    })
+
+    const template = dedent`
+    Hello world
+
+    [alert]
+    hello
+    [/alert]
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+
+    assert.equal(result.contents, dedent`
+    <p>Hello world</p>\n
+    `)
+  })
+
+  test('work fine when inline macro is the last line', async (assert) => {
+    const macroable = Macroable()
+    assert.plan(2)
+
+    macroable.addMacro('codepen', function (props, { eat }) {
+      assert.deepEqual(props, { src: 'foo' })
+    }, true)
+
+    const template = dedent`
+    Here is a pen
+
+    [codepen src=foo]
+    `
+    const result = await exec(template, unifiedStream().use(macroable.transformer).use(html))
+
+    assert.equal(result.contents, dedent`
+    <p>Here is a pen</p>\n
     `)
   })
 })