+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/poppinss/remark-macro/compare/v1.0.2...v1.0.3) (2018-03-26)
+
+
+### Bug Fixes
+
+* **prop:** prop builder now allows quoted strings ([5dd256d](https://github.com/poppinss/remark-macro/commit/5dd256d))
+
+
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/poppinss/remark-macro/compare/v1.0.1...v1.0.2) (2018-03-25)
+
+
+### Features
+
+* **macro:** allow inline macros ([b49e537](https://github.com/poppinss/remark-macro/commit/b49e537))
+
+
+
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/poppinss/remark-macro/compare/v1.0.0...v1.0.1) (2018-03-24)
+
+
+### Bug Fixes
+
+* **transformer:** fix typo in key name transformer ([bd49b77](https://github.com/poppinss/remark-macro/commit/bd49b77))
+
+
+### Features
+
+* **macro:** trim macro content by default ([842090d](https://github.com/poppinss/remark-macro/commit/842090d))
+
+
+
 <a name="1.0.0"></a>
 # 1.0.0 (2018-03-23)